requirements.txt
fastapi
uvicorn
python-docx
python-dotenv
pydantic